Move styles to general.ts and replace UI for contact support form
