+++ components/brave_vpn/resources/panel/components/general.ts
